requirements.txt
# Core Requirements
Flask==2.3.2
Werkzeug==2.3.7
Jinja2==3.1.2

# Database
mysql-connector-python==8.1.0
SQLAlchemy==2.0.19  # Optional but recommended for some utilities

# Security
bcrypt==4.0.1
PyJWT==2.8.0

# Form Handling
WTForms==3.0.1
Flask-WTF==1.1.1

# API Support
Flask-RESTful==0.3.10
Flask-CORS==4.0.0

# Authentication
Flask-Login==0.6.2
Flask-Session==0.5.0

# Utilities
python-dotenv==1.0.0
pytz==2023.3
python-dateutil==2.8.2

# Development
Flask-DebugToolbar==0.13.1
Faker==19.6.2
pytest==7.4.0
pytest-cov==4.1.0

# Frontend Assets (if needed)
Flask-Assets==2.0
cssmin==0.2.0
jsmin==3.0.1

# Email (if needed)
Flask-Mail==0.9.1